phase3-django-api/lib_mng_sys/pyproject.toml
[project]
name = "lib-mng-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.6",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-rest-framework>=0.1.0",
    "pymysql>=1.1.1",
    "python-decouple>=3.8",
]
